requirements.txt
# Media Pipeline Dependencies
# Core dependencies
PyYAML>=6.0
Pillow>=9.0.0
python-dateutil>=2.8.0
requests>=2.28.0

# iCloud integration
pyicloud>=0.10.0

# Web UI dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0
python-socketio>=5.8.0
eventlet>=0.33.0

# System monitoring
psutil>=5.9.0

# Telegram integration
python-telegram-bot>=20.0

# Enhanced logging
colorlog>=6.7.0

# Scheduling
schedule>=1.2.0

# Database (SQLite is built-in to Python)
# No additional database dependencies needed

# Development dependencies (uncomment if needed)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
